scanner/requirements.txt
Pillow>=10.0.0
imagehash>=4.3.1
requests>=2.31.0
watchdog>=3.0.0
numpy>=1.24.0
opencv-python>=4.8.0
python-dotenv>=1.0.0
